refactor: PetFood 필드 일급컬렉션 적용
